Move type parsing functions to separate file.
