+++ src/model/index.js
@@ -6,7 +6,8 @@ import mergeEntity from './merge.js';
 import updateEntity from './update.js';
 import defaultEntities, { entityMethods } from '../entities.js';
 import { dereference } from '../json-schema/reference.js';
-import { createObserver, isObject, parseEntityType } from '../utils.js';
+import { parseEntityType } from '../types.js';
+import { createObserver, isObject } from '../utils.js';
 
 /**
  * JSON Schema for defining the entities supported by a farmOS instance.
